# [julia] add comments, reduce `getindex()` hot spot, use MVector to speed up