@algunion my local benchmark shows

## Before
```julia
BenchmarkTools.Trial: 170 samples with 1 evaluation.
 Range (min … max):  29.034 ms …  32.120 ms  ┊ GC (min … max): 0.00% … 4.70%
 Time  (median):     29.381 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   29.506 ms ± 397.714 μs  ┊ GC (mean ± σ):  0.05% ± 0.44%

     ▃ ▁▂ ▄  █▃▂     ▁
  ▄▁▁█▃█████████▆▆▆█▆█▃▃▁▅▅▄▄▁▁▅▅▅▃▁▄▁▁▅▁▃▁▁▃▁▁▃▃▁▃▃▁▃▁▃▁▁▁▃▁▃ ▃
  29 ms           Histogram: frequency by time         30.7 ms <

 Memory estimate: 1.29 MiB, allocs estimate: 181.
```

## This PR
```julia
BenchmarkTools.Trial: 257 samples with 1 evaluation.
 Range (min … max):  19.041 ms …  23.469 ms  ┊ GC (min … max): 0.00% … 13.66%
 Time  (median):     19.400 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   19.493 ms ± 476.229 μs  ┊ GC (mean ± σ):  0.12% ±  1.02%

    ▁█▂▃▅ ▄▃▁
  ▅██████████▆▅▃▇▆▄▄▃▃▃▁▁▃▁▃▂▁▃▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂ ▃
  19 ms           Histogram: frequency by time         21.8 ms <

 Memory estimate: 1.29 MiB, allocs estimate: 181.
 ```